Update (More in description)
Changed solidity compiler to 0.8.19 to be able to use some methods, added functions to market.js, set persistence when entering the market. Updated contract with more functions to get and create random NFTs, maybe will switch to images, I just used random strings for testing. The functions in the market.js has to be binded later into the three buttons in the frontend.