Big Update (See Description)
Ok so, I have decided that generating a random prompt is impossibile, hence the items will be listed in the json files (instead of using a database), each time an item is generated it gets removed from the list and added to the used one to avoid repetition. Added a quest system logic that follows the same rules kinda. The only thing is I didn't add any logic for the smart contract so it's just 'server operations' for now, but it works.
And of corse it needs a lot of polish.