Modified DB
Now the quests have both company name and address, so it is easier to read when later doing the transaction. However companies.db is still necessary cause it is suited when company X wants to add another quest. Only thing is that the quest is the same among the three companies.