Updated NFTs
Made the modal to get the nfts, added retrivial of the image through the server, removed rarirty from contract and handle it from the backend instead, added nft database in script.

To Do: modify the get of NFT's in the backend to maybe do checks on the smart contract first to ensure consistency, and then to add the trading option in the nft modal to set your nft to be tradable in the market.