New NFTs design
Now all the information of the NFT (without the description) is stored in the
contract so the nfts database is no longer required.

Added in the contract the User struct wich memorize all the NFTs of a
particular user so it is easy to recover them. The NFT in the market are
stored in one array.

Next Step:
- implement the parsing of the all NFT in the frontend and
  start implementing the logic of the market
- Add some checks in the contract for security reason (non mandatory)