NFTs Update
Nfts are reconstructed and displayed correctly both in the modal and in the market.

The only thing is that to set the nft as 'to be sold' and viceversa, I added a field in the contract and two functions. But to set the nfts 'to be sold' still requires a cost, so maybe there should be a warning that tells you that selling the nft has a small price to be payed.

Lastly we still have to think on how to decide for the nft price, and how to actually sell it.