Merge the NFT handler with the Market class
Write the market function to sell, unsell and buy the NFTs. Need to be
tested yet. The last thing to do is to write the socket communication
that must update the user's NFTs and the market after one of the
previous operations is executed.

For now the cost of one NFT is fixed to 2 ETH, maybe a strategy to
decide the NFT value need to be implemented.